Planet popups now always display health, and also display fleet strength.
This commit is a refinement of a system originally developed by Skeletonxf.

------------------------------
Note that fleet strength might still count as zero if planetary defenses are online but no support ships exist. This is because strength is calculated by multiplying fleet HP and DPS, and a planet can't be damaged by conventional weapons.

More testing is required on that point.