ProduceNativePressurePct will now produce pressure for all native resources and update when the resources change
A change in the number of native resources, the amount of pressure percent to produce (as before), or every 30 seconds, will trigger a recalculation. This ensures First buildings will produce native pressure of a resource added after the buildings, and the 30 second recalculation covers the edge case in which a planet is terraformed, changing the resource, but not the number of them. Now, even resource changing will update the pressure percentage produced within 30 seconds in the worst case.

It might be worth complementing this with a hook put onto every terraforming method that triggers a refresh of this calculation, but as this currently is, the 30 second checks will trigger a pressure update when everything else fails to.