Make mothership pop go away when no longer a mothership
Fixes huge exploit where players could build loads of motherships and retrofit them to size 16 2 hex ships and retain the population/income. Ports the hooks from CE that I used to do this, so also adds a convenient way to make mothership pop vary (as in CE), but for the modpack obviously mothership pop is either 16 with the hull or 0 without.