Attack orders quality of life changes ported from colonisation expansion
Of particular note to point out, I opted to not add an special case handling for ships owned by empires that are ForbidDeepSpace, firstly because you can't make an empire ForbidDeepSpace in vanilla anyway, and any particulars about how flux is modded in is beyond the remit of the modpack at least for as long as it doesn't add flux support, and secondly, because the MoveTo order will happily send a ship into deep space anyway, so it is more consistent for Chase and Attack to disregard the limitation as well. If someone wants to port these orders to behave differently with Flux, I suggest chase is modified to abort if the target flees (or perhaps to flux after the target or something??) and keep distance is modified to clip to evasion point to within the current region.

I've also ported the attack context menu changes mentioned in https://github.com/sol-oriens/SR2-Community-Patch/issues/6, so planets can now be ordered to attack as well.

Vanilla has some odd defaults with how battleworlds behave, which I've not changed because they are absolutely game play changing and not really QoL things, but as a heads up: planets that fly themselves into deep space cannot be annexed or captured while they stay in deep space. Planets that fly into deep space even momentarily also grant vision of themselves to everyone and this vision lasts for the rest of the game. Once vision of a planet is obtained, it continues to show the current position of the planet - this is why you always know where a planet is when it is orbiting a star, but it also makes battleworlds in vanilla incapable of hiding. The QoL changes to battleworlds should at least make them usable, but the mechanics blind mind chose are interesting nonetheless.

Although I've made the capture order better at entering orbit of the target planet, it still won't capture a planet in deep space, because of the vanilla mechanics explained above.