Added "social media preview image" for GitHub.
The other OpenSR repos got similar images, but I haven't decided whether I should push them to GitHub yet. Recreating them is a trivial matter of replacing the tagline and rerunning the drop shadow effect on the text layer, so I could be convinced of the value of either approach.